Somsak source code version 2.1.9.7 on 2018-10-23

+++ src/ntu/celt/eUreka2/pages/modules/assessment/Manage.java
@@ -203,8 +203,8 @@ public class Manage extends AbstractPageAssessment{
 			return true;
 		return false;
 	}
-	public int getTotalWeight(){
-		int total = 0;
+	public float getTotalWeight(){
+		float total = 0;
 		for(Assessment asst: assmts){
 			total += asst.getWeightage();
 		}
